# dockstore-webservice/src/main/resources/openapi3/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: theglobalalliance@genomicsandhealth.org
    name: Dockstore@ga4gh
    url: https://discuss.dockstore.org/t/opening-helpdesk-tickets/1506
  description: "This describes the dockstore API, a webservice that manages pairs\
    \ of Docker images and associated metadata such as CWL documents and Dockerfiles\
    \ used to build those images. Explore swagger.json for a Swagger 2.0 description\
    \ of our API and explore openapi.yaml for OpenAPI 3.0 descriptions."
  license:
    name: Apache License Version 2.0
    url: https://github.com/dockstore/dockstore/blob/develop/LICENSE
  termsOfService: TBD
  title: Dockstore API
  version: 1.11.0-alpha.2-SNAPSHOT
servers:
- description: Current server when hosted on AWS
  url: /api
  variables: {}
- description: When working locally
  url: /
  variables: {}
- description: Production server
  url: https://dockstore.org/api
  variables: {}
- description: Staging server
  url: https://staging.dockstore.org/api
  variables: {}
- description: Nightly build server
  url: https://dev.dockstore.net/api
  variables: {}
...
paths:
...
  /aliases/workflow-versions/{workflowVersionId}:
    post:
      description: Add aliases linked to a workflow version in Dockstore.
      operationId: addAliases
      parameters:
      - in: path
        name: workflowVersionId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: aliases
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowVersion'
          description: default response
      security:
      - bearer: []
      tags:
      - aliases
...
  /entries/{entryId}/exportToOrcid:
    post:
      description: Export entry to ORCID. DOI is required
      operationId: exportToORCID
      parameters:
      - description: The id of the entry to export.
        in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - description: Optional version ID of the entry version to export.
        in: query
        name: versionId
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearer: []
      tags:
      - entries
  /entries/{entryId}/verifiedPlatforms:
    get:
      description: Get the verified platforms for each version of an entry.
      operationId: getVerifiedPlatforms
      parameters:
      - description: id of the entry
        in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VersionVerifiedPlatform'
          description: default response
      security:
      - bearer: []
      tags:
      - entries
  /entries/{entryId}/versions/{versionId}/fileTypes:
    get:
      description: Retrieve the unique file types of a version's sourcefile
      operationId: getVersionsFileTypes
      parameters:
      - description: Entry to retrieve the version from
        in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - description: Version to retrieve the sourcefile types from
        in: path
        name: versionId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  enum:
                  - DOCKSTORE_CWL
                  - DOCKSTORE_WDL
                  - DOCKERFILE
                  - CWL_TEST_JSON
                  - WDL_TEST_JSON
                  - NEXTFLOW
                  - NEXTFLOW_CONFIG
                  - NEXTFLOW_TEST_PARAMS
                  - DOCKSTORE_YML
                  - DOCKSTORE_SERVICE_YML
                  - DOCKSTORE_SERVICE_TEST_JSON
                  - DOCKSTORE_SERVICE_OTHER
                  - DOCKSTORE_GXFORMAT2
                  - GXFORMAT2_TEST_FILE
                  - DOCKSTORE_SWL
                  - SWL_TEST_JSON
                uniqueItems: true
          description: default response
      security:
      - bearer: []
      tags:
      - entries
  /entries/{id}/aliases:
    post:
      description: Add aliases linked to a entry in Dockstore.
      operationId: addAliases_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: aliases
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: Successfully added alias to entry
      security:
      - bearer: []
      tags:
      - entries
  /entries/{id}/collections:
    get:
      description: Get the collections and approved organizations that contain the
        published entry
      operationId: entryCollections
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionOrganization'
          description: default response
      tags:
      - entries
  /entries/{id}/topic:
    post:
      description: Create a discourse topic for an entry.
      operationId: setDiscourseTopic
      parameters:
      - description: The id of the entry to add a topic to.
        in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: default response
      security:
      - bearer: []
      tags:
      - entries
...
  /organizations/{organizationId}/collections/{collectionId}/entry:
    delete:
      description: Delete an entry to a collection.
      operationId: deleteEntryFromCollection
      parameters:
      - description: Organization ID.
        in: path
        name: organizationId
        required: true
        schema:
          type: integer
          format: int64
      - description: Collection ID.
        in: path
        name: collectionId
        required: true
        schema:
          type: integer
          format: int64
      - description: Entry ID.
        in: query
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - description: Version ID.
        in: query
        name: versionId
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
          description: default response
      security:
      - bearer: []
      summary: Delete an entry to a collection.
      tags:
      - organizations
    post:
      description: Add an entry to a collection.
      operationId: addEntryToCollection
      parameters:
      - description: Organization ID.
        in: path
        name: organizationId
        required: true
        schema:
          type: integer
          format: int64
      - description: Collection ID.
        in: path
        name: collectionId
        required: true
        schema:
          type: integer
          format: int64
      - description: Entry ID.
        in: query
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - description: Version ID.
        in: query
        name: versionId
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
          description: default response
      security:
      - bearer: []
      summary: Add an entry to a collection.
      tags:
      - organizations
...
  /workflows/path/entry/{repository}:
    get:
      description: Get an entry by path.
      operationId: getEntryByPath
      parameters:
      - in: path
        name: repository
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: default response
      security:
      - bearer: []
      tags:
      - workflows
...
  /workflows/{alias}/aliases:
    get:
      description: Retrieves a workflow by alias.
      operationId: getWorkflowByAlias
      parameters:
      - in: path
        name: alias
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
          description: default response
      security:
      - bearer: []
      tags:
      - workflows
...
components:
  schemas:
    Alias:
      type: object
      properties:
        content:
          type: string
...
    CollectionOrganization:
      type: object
      properties:
        collectionDisplayName:
          type: string
        collectionId:
          type: integer
          format: int64
        collectionName:
          type: string
        organizationDisplayName:
          type: string
        organizationId:
          type: integer
          format: int64
        organizationName:
          type: string
...
    FileFormat:
      type: object
      properties:
        value:
          type: string
...
    VersionMetadata:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orcidPutCode:
          type: string
        parsedInformationSet:
          type: array
          items:
            $ref: '#/components/schemas/ParsedInformation'
    VersionVerifiedPlatform:
      type: object
      properties:
        metadata:
          type: string
        path:
          type: string
        platformVersion:
          type: string
        source:
          type: string
        verified:
          type: boolean
        versionId:
          type: integer
          format: int64
...
    WorkflowVersion:
      type: object
      properties:
        aliases:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Alias'
        author:
          type: string
        commitID:
          type: string
        dbUpdateDate:
          type: integer
          format: int64
        description:
          type: string
        descriptionSource:
          type: string
          enum:
          - README
          - DESCRIPTOR
        dirtyBit:
          type: boolean
        doiStatus:
          type: string
          enum:
          - NOT_REQUESTED
          - REQUESTED
          - CREATED
        doiURL:
          type: string
        email:
          type: string
        frozen:
          type: boolean
        hidden:
          type: boolean
        id:
          type: integer
          format: int64
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
          uniqueItems: true
        input_file_formats:
          type: array
          items:
            $ref: '#/components/schemas/FileFormat'
          uniqueItems: true
        last_modified:
          type: integer
          format: int64
        legacyVersion:
          type: boolean
        name:
          type: string
          description: "Implementation specific, can be a quay.io or docker hub tag\
            \ name"
          example: latest
        output_file_formats:
          type: array
          items:
            $ref: '#/components/schemas/FileFormat'
          uniqueItems: true
        parent:
          $ref: '#/components/schemas/EntryObjectObject'
        reference:
          type: string
          description: git commit/tag/branch
          example: master
        referenceType:
          type: string
          enum:
          - COMMIT
          - TAG
          - BRANCH
          - NOT_APPLICABLE
          - UNSET
        subClass:
          type: string
          enum:
          - DOCKER_COMPOSE
          - SWARM
          - KUBERNETES
          - HELM
        synced:
          type: boolean
        valid:
          type: boolean
        validations:
          type: array
          items:
            $ref: '#/components/schemas/Validation'
          uniqueItems: true
        verified:
          type: boolean
        verifiedSource:
          type: string
        verifiedSources:
          type: array
          items:
            type: string
        versionEditor:
          $ref: '#/components/schemas/User'
        versionMetadata:
          $ref: '#/components/schemas/VersionMetadata'
        workflow_path:
          type: string
        workingDirectory:
          type: string
      required:
      - name
      - reference
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http